--- a/test/localstack.ts
+++ b/test/localstack.ts
@@ -1,3 +1,4 @@
+import { performance, PerformanceObserver } from "node:perf_hooks";
 import * as cxapi from "@aws-cdk/cx-api";
 import { App, CfnOutput, Stack } from "aws-cdk-lib";
 import { SdkProvider } from "aws-cdk/lib/api/aws-auth";
@@ -21,7 +22,10 @@ export const clientConfig = {
 const CF = new CloudFormation(clientConfig);
 
 export const deployStack = async (app: App, stack: Stack) => {
-  const cloudAssembly = await asyncSynth(app);
+  performance.mark("AwaitSynth");
+  const cloudAssembly = await asyncSynth(app, {});
+  performance.measure("After await synth", "AwaitSynth");
+  performance.measure("Add resources", "InvokeResourceGetter");
 
   const sdkProvider = await SdkProvider.withAwsCliCompatibleDefaults({
     httpOptions: clientConfig as any,
@@ -38,21 +42,29 @@ export const deployStack = async (app: App, stack: Stack) => {
     credentials: clientConfig.credentials,
   };
 
+  performance.mark("CfnCompile");
+
   const cfn = new CloudFormationDeployments({
     sdkProvider,
   });
 
-  const assembly = new cxapi.CloudAssembly(cloudAssembly.directory);
-  const stackArtifact = cxapi.CloudFormationStackArtifact.fromManifest(
-    assembly,
-    stack.artifactId,
-    cloudAssembly.getStackArtifact(stack.artifactId).manifest
-  ) as cxapi.CloudFormationStackArtifact;
+  // const assembly = new cxapi.CloudAssembly(cloudAssembly.directory);
+  // const stackArtifact = cxapi.CloudFormationStackArtifact.fromManifest(
+  //   cloudAssembly,
+  //   stack.artifactId,
+  //   cloudAssembly.getStackArtifact(stack.artifactId).manifest
+  // ) as cxapi.CloudFormationStackArtifact;
 
+  performance.measure("Complete CfnCompile", "CfnCompile");
+
+  performance.mark("DeployStack");
   await cfn.deployStack({
-    stack: stackArtifact,
+    stack: cloudAssembly.getStackArtifact(
+      stack.artifactId
+    ) as unknown as cxapi.CloudFormationStackArtifact,
     force: true,
   });
+  performance.measure("Deploy Stack complete", "DeployStack");
 };
 
 interface ResourceReference<Outputs extends Record<string, string>> {
@@ -105,7 +117,19 @@ export const localstackTestSuite = (
 
   let stackOutputs: CloudFormation.Outputs | undefined;
 
+  const obs = new PerformanceObserver((items) => {
+    items
+      .getEntries()
+      .forEach(({ duration, startTime, name, entryType }) =>
+        console.log(entryType, name, startTime, duration)
+      );
+  });
+  obs.observe({ entryTypes: ["mark", "measure", "function"], buffered: true });
+  performance.measure("Start to Now");
+
   beforeAll(async () => {
+    performance.mark("Before");
+
     testContexts = tests.map(({ resources, skip }, i) => {
       // create the construct on skip to reduce output changes when moving between skip and not skip
       const construct = new Construct(stack, `parent${i}`);
@@ -128,11 +152,18 @@ export const localstackTestSuite = (
       return {};
     });
 
-    await deployStack(app, stack);
+    performance.mark("DeployStack");
+    await performance.timerify(deployStack)(app, stack);
+    performance.measure("After deploy stack", "DeployStack");
+
+    performance.mark("DescribeStacks");
 
     stackOutputs = (
       await CF.describeStacks({ StackName: stack.stackName }).promise()
     ).Stacks?.[0].Outputs;
+
+    performance.measure("After describe stacks", "DescribeStacks");
+    performance.mark("BeforeEnd");
   });
 
   // @ts-ignore
@@ -144,8 +175,11 @@ export const localstackTestSuite = (
   };
 
   // register tests
+  performance.mark("GetTests");
   fn(testResource, stack, app);
+  performance.measure("After getting tests", "GetTests");
 
+  performance.mark("InvokeResourceGetter");
   tests.forEach(({ name, test: testFunc, skip }, i) => {
     if (!skip) {
       test(name, () => {
@@ -164,4 +198,6 @@ export const localstackTestSuite = (
       test.skip(name, () => {});
     }
   });
+  performance.mark("SyncResource");
+  performance.measure("Sync resource time", "SyncResource");
 };