Improving the performance of the localstack tests.

Currently seeing 240s to fun the `function.localstack` tests with a 12m build time.

Current Improvements:
* [x] Simplify the stack deployment logic saw a 1+ minute improvement on function deployment, hoping for 1-3 minute overall improvement.
   * [x] Clean up
* [x] Write up detailed performance breakdown
* [x] Remove the performance measuring code added


* tsc + functionless - 50 seconds
* Add CDK, bundle, and run serializer - 160 seconds
	* Bundle?
* ~~CdnCompile - 80 seconds~~
* Deploy Stack - 30 seconds