@sam-goodwin 2022-06-02
Build failed after 13 minutes. Sure it's an optimization?