@thantos 2022-06-02
## Small Writeup

CDK exports two libraries, `@aws-cdk/cx-api` and `aws-cdk-lib/cx-api` 

the Stack synth produces `aws-cdk-lib/cx-api` CloudArtifact, but to deploy programmatically, we needed  `@aws-cdk/cx-api` CloudArtifact.

To do this, (copied code from a github issue) we load the assembly output from `aws-cdk-lib/cx-api` into a new  `@aws-cdk/cx-api`.

This worked, but was taking 80+ seconds.

Before

```ts
// synth
const cloudAssembly = await asyncSynth(app);
const sdkProvider = await SdkProvider.withAwsCliCompatibleDefaults({
  httpOptions: clientConfig as any,
});

// config hacking...
// @ts-ignore
sdkProvider.sdkOptions = {
  // @ts-ignore
  ...sdkProvider.sdkOptions,
  endpoint: clientConfig.endpoint,
  s3ForcePathStyle: clientConfig.s3ForcePathStyle,
  accessKeyId: clientConfig.credentials.accessKeyId,
  secretAccessKey: clientConfig.credentials.secretAccessKey,
  credentials: clientConfig.credentials,
};

// convert the aws-cdk-lib/cx-api CloudAssembly into a @aws-cdk/cx-api CloudAssembly
const assembly = new cxapi.CloudAssembly(cloudAssembly.directory);
const stackArtifact = cxapi.CloudFormationStackArtifact.fromManifest(
	assembly,
	stack.artifactId,
	assembly.getStackArtifact(stack.artifactId).manifest
) as cxapi.CloudFormationStackArtifact

// deploy using 
await cfn.deployStack({
    stack: stackArtifact,
	force: true,
});
```

#### Deeper Dive

##### Why did we need to do this? 

Well... trying to use the output of the stack form `aws-cdk-lib` produced typescript issues.

![Pasted image 20220601141858](https://user-images.githubusercontent.com/289213/171621841-e2ca85ad-da4d-48b1-b779-91d1cd934c61.png)

##### Why are the types different?

Two different npm packages.

##### Where does the code come form?

From the same source. https://github.com/aws/aws-cdk/tree/7eda25625ec1d15fe610097b0456438d6806d9c9/packages/@aws-cdk/cx-api

##### Why are the bundles different?

[CDK Strips](https://docs.aws.amazon.com/cdk/v2/guide/migrating-v2.html) deprecated fields, methods, and classes out for the v2 release of CDK (`aws-cdk-lib`) but they do not do this for `aws-cdk/cx-api`. 

The types be coerced `cloudAssembly.getStackArtifact(stack.artifactId) as unknown as cxapi.CloudFormationStackArtifact` from the `aws-cdk-lib` type to the `aws-cdk/cx-api` form as long as neither side expect use the of deprecated fields.

After

```ts
const cloudAssembly = await asyncSynth(app);
const sdkProvider = await SdkProvider.withAwsCliCompatibleDefaults({
  httpOptions: clientConfig as any,
});

// @ts-ignore
sdkProvider.sdkOptions = {
  // @ts-ignore
  ...sdkProvider.sdkOptions,
  endpoint: clientConfig.endpoint,
  s3ForcePathStyle: clientConfig.s3ForcePathStyle,
  accessKeyId: clientConfig.credentials.accessKeyId,
  secretAccessKey: clientConfig.credentials.secretAccessKey,
  credentials: clientConfig.credentials,
};

await cfn.deployStack({
    stack: cloudAssembly.getStackArtifact(stack.artifactId) as unknown as cxapi.CloudFormationStackArtifact,
	force: true,
});
```