@thantos 2022-06-02
> Build failed after 13 minutes. Sure it's an optimization?

I THINK that jest doesn't like the performance observer and console logs which are now removed. We'll see. Its running faster in general on my machine.